First Light for the versatile duck
What I have here is close enough to correct to capture. As of this commit it's polluted with development artefacts (code snippets, old style model files...) but it runs and uses the correct architecture (Application Model -> Assets, Note that both element layout and rendering are considered code assets, not part of the framework itself.